z3fold deprecated: update the comment to indicate that
Thankfully, we have already switched to using zsmalloc.

Linux kernel 6.14 is deprecating z3fold
(and telling users to use zsmalloc instead),
so it will be removed eventually from the kernel
and should not be used any longer.

Leaving a mention of it in the documentation
since so much of the internet information
refers glowingly to z3fold...,
so mention its deprecation so no one tries to use it anymore.